move files around some more
