+++ src/renderer/pages/loading/loading_page.tsx
